+++ Steps/IO/Compression/src/CreateNewZipArchiveStep.cs
@@ -1,7 +1,7 @@
 ﻿// <copyright file="CreateNewZipArchiveStep.cs" company="Chris Trout">
 // MIT License
 //
-// Copyright(c) 2020-2021 Chris Trout
+// Copyright(c) 2020-2022 Chris Trout
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -26,6 +26,7 @@ namespace MyTrout.Pipelines.Steps.IO.Compression
 {
     using Microsoft.Extensions.Logging;
     using System;
+    using System.Collections.Generic;
     using System.Globalization;
     using System.IO;
     using System.IO.Compression;
@@ -34,7 +35,7 @@ namespace MyTrout.Pipelines.Steps.IO.Compression
     /// <summary>
     /// Unzips a <see cref="System.IO.Stream"/> and calls downstream once for each file in the zip archive.
     /// </summary>
-    public class CreateNewZipArchiveStep : AbstractPipelineStep<CreateNewZipArchiveStep>
+    public class CreateNewZipArchiveStep : AbstractCachingPipelineStep<CreateNewZipArchiveStep>
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="CreateNewZipArchiveStep" /> class with the specified parameters.
@@ -47,55 +48,41 @@ namespace MyTrout.Pipelines.Steps.IO.Compression
             // no op
         }
 
+        /// <inheritdoc />
+        public override IEnumerable<string> CachedItemNames => new List<string>() { CompressionConstants.ZIP_ARCHIVE, PipelineContextConstants.OUTPUT_STREAM };
+
         /// <summary>
         /// Unzips the stream and calls the downstream caller once for each file in the zip archive.
         /// </summary>
         /// <param name="context">The pipeline context.</param>
         /// <returns>A completed <see cref="Task" />.</returns>
-        protected override async Task InvokeCoreAsync(IPipelineContext context)
+        protected override async Task InvokeCachedCoreAsync(IPipelineContext context)
         {
-            if (context.Items.TryGetValue(CompressionConstants.ZIP_ARCHIVE, out object? previousArchive))
-            {
-                context.Items.Remove(CompressionConstants.ZIP_ARCHIVE);
-            }
-
-            if (context.Items.TryGetValue(PipelineContextConstants.OUTPUT_STREAM, out object? previousOutputStream))
-            {
-                if (previousOutputStream is IAsyncDisposable workingStream)
-                {
-                    await workingStream.DisposeAsync().ConfigureAwait(false);
-                }
-
-                context.Items.Remove(PipelineContextConstants.OUTPUT_STREAM);
-            }
-
-            var outputStream = new MemoryStream();
-
-            using (ZipArchive zipArchive = new ZipArchive(outputStream, ZipArchiveMode.Create, leaveOpen: true))
+            using (var outputStream = new MemoryStream())
             {
                 try
                 {
-                    this.Logger.LogInformation(Resources.ZIP_ARCHIVE_OPENED(CultureInfo.CurrentCulture, nameof(CreateNewZipArchiveStep)));
+                    using (ZipArchive zipArchive = new ZipArchive(outputStream, ZipArchiveMode.Create, leaveOpen: true))
+                    {
+                        this.Logger.LogInformation(Resources.ZIP_ARCHIVE_OPENED(CultureInfo.CurrentCulture, nameof(CreateNewZipArchiveStep)));
 
-                    context.Items.Add(CompressionConstants.ZIP_ARCHIVE, zipArchive);
+                        context.Items.Add(PipelineContextConstants.OUTPUT_STREAM, outputStream);
+                        context.Items.Add(CompressionConstants.ZIP_ARCHIVE, zipArchive);
 
-                    this.Logger.LogDebug(Resources.ZIP_ARCHIVE_ADDED_TO_PIPELINE(CultureInfo.CurrentCulture));
+                        this.Logger.LogDebug(Resources.ZIP_ARCHIVE_ADDED_TO_PIPELINE(CultureInfo.CurrentCulture));
 
-                    await this.Next.InvokeAsync(context).ConfigureAwait(false);
-
-                    context.Items.Add(PipelineContextConstants.OUTPUT_STREAM, outputStream);
+                        await this.Next.InvokeAsync(context).ConfigureAwait(false);
+                    }
                 }
                 finally
                 {
+                    // Not logging the removal of output stream.
+                    context.Items.Remove(PipelineContextConstants.OUTPUT_STREAM);
+
                     if (context.Items.ContainsKey(CompressionConstants.ZIP_ARCHIVE))
                     {
                         context.Items.Remove(CompressionConstants.ZIP_ARCHIVE);
                         this.Logger.LogDebug(Resources.ZIP_ARCHIVE_REMOVED_FROM_PIPELINE(CultureInfo.CurrentCulture));
-                    }
-
-                    if (previousArchive != null)
-                    {
-                        context.Items.Add(CompressionConstants.ZIP_ARCHIVE, previousArchive);
                     }
                 }
             }