+++ Steps/IO/Compression/test/RemoveItemFromContextStep.cs
@@ -1,0 +1,72 @@
+﻿// <copyright file="RemoveItemFromContextStep.cs" company="Chris Trout">
+// MIT License
+//
+// Copyright(c) 2022 Chris Trout
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+// </copyright>
+
+namespace MyTrout.Pipelines.IO.Compression.Tests
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Threading.Tasks;
+
+    [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+    internal sealed class RemoveItemFromContextStep : IPipelineRequest
+    {
+        public RemoveItemFromContextStep(params string[] contextItemsToRemove)
+        {
+            this.ContextItemsToRemove = contextItemsToRemove;
+        }
+
+        public IEnumerable<string> ContextItemsToRemove { get; init; }
+
+        /// <inheritdoc />
+        public void Dispose()
+        {
+            GC.SuppressFinalize(this);
+        }
+
+        /// <summary>
+        /// Disposes of any disposable resources for this instance.
+        /// </summary>
+        /// <returns>A completed <see cref="ValueTask" />.</returns>
+        /// <remarks>Developers who need to dispose of unmanaged resources should override this method.</remarks>
+        public ValueTask DisposeAsync()
+        {
+            // No async cleanup required.
+
+            // Dispose of synchronous unmanaged resources.
+            this.Dispose();
+
+            return default;
+        }
+
+        public Task InvokeAsync(IPipelineContext context)
+        {
+            foreach (string key in this.ContextItemsToRemove)
+            {
+                context.Items.Remove(key);
+            }
+
+            return Task.CompletedTask;
+        }
+    }
+}